requirements.txt
Package                      Version
---------------------------- ---------
annotated-types              0.7.0
anyio                        4.9.0
certifi                      2025.4.26
click                        8.2.1
dnspython                    2.7.0
email_validator              2.2.0
exceptiongroup               1.3.0
fastapi                      0.115.12
fastapi-cli                  0.0.7
flatbuffers                  25.2.10
h11                          0.16.0
h5py                         3.14.0
httpcore                     1.0.9
httptools                    0.6.4
httpx                        0.28.1
idna                         3.10
Jinja2                       3.1.6
keras                        2.11.0
libclang                     18.1.1
Markdown                     3.8
markdown-it-py               3.0.0
MarkupSafe                   3.0.2
mdurl                        0.1.2
motor                        3.7.1
numpy                        2.2.6
oauthlib                     3.2.2
opt_einsum                   3.4.0
packaging                    25.0
pillow                       11.2.1
pip                          25.1.1
protobuf                     3.19.6
pyasn1                       0.6.1
pydantic                     2.11.5
pydantic_core                2.33.2
Pygments                     2.19.1
pymongo                      4.13.0
python-dotenv                1.1.0
python-multipart             0.0.20
PyYAML                       6.0.2
rich                         14.0.0
rich-toolkit                 0.14.7
setuptools                   58.1.0
shellingham                  1.5.4
six                          1.17.0
sniffio                      1.3.1
starlette                    0.46.2
tensorboard-data-server      0.6.1
tensorboard-plugin-wit       1.8.1
tensorflow-estimator         2.11.0
tensorflow-io-gcs-filesystem 0.37.1
termcolor                    3.1.0
typer                        0.16.0
typing_extensions            4.14.0
typing-inspection            0.4.1
urllib3                      2.4.0
uvicorn                      0.34.3
uvloop                       0.21.0
watchfiles                   1.0.5
websockets                   15.0.1
Werkzeug                     3.1.3
wheel                        0.45.1
wrapt                        1.17.2
